Adding --dumsmt option to dump all SMT queries

+++ tools/halmos.sh
@@ -9,22 +9,29 @@ sig="$1"; shift
 ds_test="$1"; shift
 tout="$1"; shift
 memout="$1"; shift
+dump_smt="$1"; shift
 
-rm -f ./*.smt2
+extra_params=""
+if [[ "$dump_smt" == "1" ]]; then
+    extra_params="${extra_params} --dump-smt-queries"
+    rm -f ./*.smt2
+fi
 
-out=$(runlim --real-time-limit="${tout}" --kill-delay=2 --space-limit="${memout}" halmos --function "${fun_name}" --contract "${contract_name}" "$@" 2>&1)
+
+out=$(runlim --real-time-limit="${tout}" --kill-delay=2 --space-limit="${memout}" halmos --function "${fun_name}" --contract "${contract_name}" ${extra_params} "$@" 2>&1)
+
 
 # Check if we emitted smt2 files. If so, copy them over to a
 # directory based on the contract file & name
-shopt -s nullglob
-set -- *.smt2
-if [ "$#" -gt 0 ]; then
-  dir="halmos-smt2/${contract_file}.${contract_name}/"
-  mkdir -p "$dir"
-  mv -f ./*.smt2 "$dir/"
-fi
-
 set +x
+if [[ "$dump_smt" == "1" ]] && [[ "$out" =~ "Generating SMT" ]]; then
+    a="s/^Generating SMT queries in \\(.*\\)/\\1/"
+    outdir=$(echo "$out" | grep "Generating SMT" | sed -e "${a}")
+    dir="halmos-smt2/${contract_file}.${contract_name}/"
+    mkdir -p "$dir"
+    mv -f ${outdir}/*.smt2 "$dir/"
+    set -x
+fi
 
 if [[ $out =~ "WARNING:Halmos:Counterexample: unknown" ]]; then
   exit 1
